refactor: move genesis provider to regenesis package
