+++ packages/core-utils/package.json
@@ -46,12 +46,9 @@
   },
   "dependencies": {
     "@ethersproject/abstract-provider": "^5.4.1",
-    "@ethersproject/bignumber": "^5.5.0",
     "@ethersproject/bytes": "^5.5.0",
-    "@ethersproject/properties": "^5.5.0",
     "@ethersproject/providers": "^5.4.5",
     "chai": "^4.3.4",
-    "ethereumjs-util": "^7.1.3",
     "ethers": "^5.4.5",
     "lodash": "^4.17.21"
   }
