+++ packages/regenesis-surgery/package.json
@@ -16,6 +16,10 @@
   "devDependencies": {
     "@discoveryjs/json-ext": "^0.5.3",
     "@eth-optimism/core-utils": "^0.6.0",
+    "@ethersproject/abstract-provider": "^5.5.1",
+    "@ethersproject/bignumber": "^5.5.0",
+    "@ethersproject/properties": "^5.5.0",
+    "@ethersproject/providers": "^5.5.0",
     "@types/node": "^15.12.2",
     "@types/node-fetch": "^3.0.3",
     "@uniswap/sdk-core": "^3.0.1",
@@ -24,13 +28,13 @@
     "byline": "^5.0.0",
     "chai-as-promised": "^7.1.1",
     "dotenv": "^10.0.0",
-    "ethereumjs-util": "^7.1.2",
+    "ethereumjs-util": "^7.1.3",
     "ethers": "^5.4.5",
     "hardhat": "^2.6.5",
+    "lint-staged": "11.0.0",
     "mocha": "^9.1.2",
     "node-fetch": "2.6.5",
     "solc": "0.8.7-fixed",
-    "ts-node": "^10.0.0",
-    "lint-staged": "11.0.0"
+    "ts-node": "^10.0.0"
   }
 }
